fix: moved vitest matcher setup from fuel-gauge to utils
