--- packages/fuel-gauge/global.d.ts
@@ -1,15 +1,0 @@
-import type { BNInput } from 'fuels';
-
-declare global {
-  namespace Vitest {
-    interface Assertion {
-      toEqualBn(expected: BNInput): void;
-    }
-    interface ExpectStatic {
-      toEqualBn(expected: BNInput): {
-        asymmetricMatch(actual: BNInput): boolean;
-        toString(): string;
-      };
-    }
-  }
-}
