--- packages/fuel-gauge/vitest.setup.ts
@@ -1,3 +1,0 @@
-import { setupTestMatchers } from './src/abi/vitest.matcher';
-
-setupTestMatchers();
