+++ packages/utils/src/test-utils.ts
@@ -2,3 +2,4 @@ export * from './test-utils/getForcProject';
 export * from './test-utils/expectToBeInRange';
 export * from './test-utils/constants';
 export * from './test-utils/wait-until-unreachable';
+export * from './test-utils/vitest.matcher';
