+++ packages/utils/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "./dist"
   },
-  "include": ["src", "test"]
+  "include": ["src", "test", "global.d.ts"]
 }
