+++ packages/utils/vitest.setup.ts
@@ -1,0 +1,3 @@
+import { setupTestMatchers } from './src/test-utils';
+
+setupTestMatchers();
